+++ packages/remix-image-wasm/src/wasmTransformer.ts
@@ -1,30 +1,45 @@
-import { initResize } from "@jsquash/resize";
 import { ImagePosition, MimeType, Transformer } from "remix-image";
 import { AvifHandler, WebpHandler, JpegHandler, PngHandler } from "./handlers";
 import { blurImage } from "./operations/blur";
 import { cropImage } from "./operations/crop";
 import { flipImage } from "./operations/flip";
-import { resizeImage } from "./operations/resize";
+import { resizeImage } from "./operations/js-resize";
 import { rotateImage } from "./operations/rotate";
 import { ImageHandler } from "./types/transformer";
 
-export const supportedInputs = new Set([
-  MimeType.JPEG,
-  MimeType.PNG,
-  MimeType.WEBP,
-  MimeType.AVIF,
-]);
+export const supportedInputs = new Set<MimeType>(
+  Object.entries({
+    [MimeType.PNG]: typeof PNG_WASM != "undefined",
+    [MimeType.JPEG]: typeof JPEG_DEC_WASM != "undefined",
+    [MimeType.WEBP]: typeof WEBP_DEC_WASM != "undefined",
+    [MimeType.AVIF]: typeof AVIF_DEC_WASM != "undefined",
+  }).reduce<MimeType[]>((accum, [mimeType, isSupported]) => {
+    if (isSupported) {
+      accum.push(mimeType as MimeType);
+    }
 
-export const supportedOutputs = new Set([
-  MimeType.JPEG,
-  MimeType.PNG,
-  MimeType.WEBP,
-  MimeType.AVIF,
-]);
+    return accum;
+  }, [])
+);
+
+export const supportedOutputs = new Set<MimeType>(
+  Object.entries({
+    [MimeType.PNG]: typeof PNG_WASM != "undefined",
+    [MimeType.JPEG]: typeof JPEG_ENC_WASM != "undefined",
+    [MimeType.WEBP]: typeof WEBP_ENC_WASM != "undefined",
+    [MimeType.AVIF]: typeof AVIF_ENC_WASM != "undefined",
+  }).reduce<MimeType[]>((accum, [mimeType, isSupported]) => {
+    if (isSupported) {
+      accum.push(mimeType as MimeType);
+    }
+
+    return accum;
+  }, [])
+);
 
 const typeHandlers: Record<string, ImageHandler> = {
-  [MimeType.JPEG]: JpegHandler,
   [MimeType.PNG]: PngHandler,
+  [MimeType.JPEG]: JpegHandler,
   [MimeType.AVIF]: AvifHandler,
   [MimeType.WEBP]: WebpHandler,
 };
@@ -33,10 +48,11 @@ export const wasmTransformer: Transformer = {
   name: "wasmTransformer",
   supportedInputs,
   supportedOutputs,
+  fallbackOutput: Array.from(supportedOutputs.values())[0],
   transform: async (
     { data, contentType: inputContentType },
     {
-      contentType: outputContentType,
+      contentType: outputContentType = inputContentType,
       width,
       height,
       fit,
@@ -52,10 +68,7 @@ export const wasmTransformer: Transformer = {
       compressionLevel,
     }
   ) => {
-    await initResize(RESIZE_WASM);
-
-    const inputHandler = typeHandlers[inputContentType];
-    let image = await inputHandler.decode(data);
+    let image = await typeHandlers[inputContentType].decode(data);
 
     if (crop) {
       image = await cropImage(image, crop, background);
@@ -86,8 +99,7 @@ export const wasmTransformer: Transformer = {
       image = await blurImage(image, blurRadius);
     }
 
-    const outputHandler = typeHandlers[outputContentType || inputContentType];
-    const result = await outputHandler.encode(image, {
+    const result = await typeHandlers[outputContentType].encode(image, {
       width: image.width,
       height: image.height,
       fit,